requirements.txt
# Dependencias principales para Machine Learning
numpy>=1.24.3
pandas>=2.0.3
matplotlib>=3.7.2
seaborn>=0.12.2
scikit-learn>=1.3.0
xgboost>=1.7.6
imbalanced-learn>=0.13.0

# Dependencias para la aplicación web
Flask>=3.0.0
Werkzeug>=3.0.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
itsdangerous>=2.2.0
click>=8.1.0

# Utilidades adicionales
joblib>=1.3.0
threadpoolctl>=3.2.0
flask>=2.3.3
flask-cors>=4.0.0
werkzeug>=2.3.7
jinja2>=3.1.2

# Dependencias adicionales
jupyter>=1.0.0
ipykernel>=6.29.5
jupyter_client          8.6.3
jupyter_core            5.8.1
kagglehub               0.3.12
kiwisolver              1.4.8
matplotlib              3.10.3
matplotlib-inline       0.1.7
nest-asyncio            1.6.0
numpy                   2.3.1
packaging               25.0
pandas                  2.3.0
parso                   0.8.4
pillow                  11.3.0
pip                     25.1.1
platformdirs            4.3.8
prompt_toolkit          3.0.51
psutil                  7.0.0
pure_eval               0.2.3
Pygments                2.19.2
pyparsing               3.2.3
python-dateutil         2.9.0.post0
pytz                    2025.2
pywin32                 310
PyYAML                  6.0.2
pyzmq                   27.0.0
requests                2.32.4
scikit-learn            1.6.1
scipy                   1.16.0
seaborn                 0.13.2
six                     1.17.0
sklearn-compat          0.1.3
stack-data              0.6.3
threadpoolctl           3.6.0
tornado                 6.5.1
tqdm                    4.67.1
traitlets               5.14.3
tzdata                  2025.2
urllib3                 2.5.0
wcwidth                 0.2.13
xgboost                 3.0.2